backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
bcrypt==4.1.2
Pillow==10.1.0
django-filter==23.5
djangorestframework-simplejwt==5.3.0
redis==5.0.1
celery==5.3.4
requests==2.31.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
lxml==4.9.3
html5lib==1.1
setuptools>=65.0.0
